// queries/src/types.ts
export type GlobalStakingInfo = {
	snxPrice: Wei;
	totalIssuedSynths: Wei;
	issuanceRatio: Wei;
	totalSupply: Wei;
	lockedSupply: Wei;
	lockedValue: Wei;
	snxStaked: Wei;
	snxPercentLocked: Wei;
	activeCRatio: Wei;
	lastDebtLedgerEntry: Wei;
};